timer refactor and interval basic setup
