+++ @internal/ssr/package.json
@@ -19,6 +19,7 @@
     "@internal/dom": "^0.0.4-alpha.0",
     "@internal/infra": "^0.0.4-alpha.0",
     "@internal/is": "^0.0.4-alpha.0",
+    "@internal/oop": "^0.0.4-alpha.0",
     "jsdom": "^19.0.0"
   },
   "devDependencies": {
