+++ @internal/ssr/src/classes/base/Scope.js
@@ -1,0 +1,10 @@
+import { abstract, abstractMethod } from '@internal/oop';
+
+const Scope = abstract(class {
+  constructor() {
+    abstractMethod(this, 'enable');
+    abstractMethod(this, 'disable');
+  }
+});
+
+export { Scope };
