+++ @internal/ssr/src/classes/base/__tests__/DigestibleScope.spec.js
@@ -1,0 +1,21 @@
+import { Scope } from '../Scope';
+
+import { DigestibleScope } from '../DigestibleScope';
+
+describe('DigestibleScope', () => {
+  it('should be a class', () => {
+    expect(DigestibleScope).toBeAClass();
+  });
+
+  it('should extend Scope', () => {
+    expect(DigestibleScope).toExtend(Scope);
+  });
+
+  it('should have an abstract method digest', () => {
+    class NoDigestScope extends DigestibleScope {
+      enable() {}
+      disable() {}
+    }
+    expect(NoDigestScope).toHaveAbstractMethod('digest');
+  });
+});
