+++ @internal/ssr/src/classes/time/Timer.js
@@ -1,0 +1,18 @@
+class Timer {
+  elapsed = 0;
+
+  get remaining() {
+    return this.wait - this.elapsed;
+  }
+
+  constructor(fn, wait) {
+    this.fn = fn;
+    this.wait = wait || 0;
+  }
+
+  tick() {
+    this.elapsed++;
+  }
+}
+
+export { Timer };
